Moving the hddm utility programs from Analysis/hddm to Utilities/hddm. Also, moving the XML (and XSD) source files from the hddm directory into the libraries/HDDM directory. The Makefile is updated to automatically generate the hddm_s.* files now if any of the event.xml, hits.xml, or reaction.xml files changes.
Note that this modification of the Makefile (actually Makefile.static)
is in contrast with a policy adopted a few years ago that stated the
schema file was the definitive source of the data model and the XML
file should be derived from that. In practice, we have seen that the
XML files are what get edited, hence this implementation. This may
be changed in the future to use the schema. In that case, the XML files
should probably be removed from the repository.

+++ libraries/HDDM/hddm_s.h
@@ -267,7 +267,7 @@ typedef struct {
 
 typedef struct {
    float                E;
-   float                dEdx;
+   double               dEdx;
    float                dradius;
    bool_t               primary;
    int32_t              ptype;
